% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BullsEyeR.R
\docType{package}
\name{BullsEyeR}
\alias{BullsEyeR}
\alias{BullsEyeR-package}
\title{Topic Modelling for Content curation     \cr
Cognizant CDB-AIM-BAI-Business Analytics \cr}
\description{
This Package provides three categories of important functions:
frequency Analysis of word tokens, Creation of Document Term Matrix and Topic Modelling using LDA.
}
\section{FreqAnalysis()}{

Frequency Analysis of word tokens - returns dataframe with words and their frequencies after initial preprocessing, sparsity control and TFIDF analysis is performed.we can pick some words from the high frequency list as custom stop words
}

\section{createDTM()}{

Creation of Document Term Matrix -repeats first step, now including the custom stop words as well, removes empty documents if any and returns a Document term matrix. This DTM is used for finding optimal number of topics for LDA modelling  using 'FindTopicsNumber' from 'ldatuning' package
}

\section{BullsEye()}{

Topic Modelling- Performs preprocessing along with removal of custom stop words,Uses topic number selected using 'ldatuning' and builds unigram topic model with/without stemming. Returns,
}

\section{Keywords}{

A data frame with keywords/dominant words of each document along with the topic number assigned by LDA model
}

\section{EmptyRows}{

A list of zero length documents after preprocessing
}

\section{Topics}{

A data frame with top 20 terms in all the topics discovered by LDA.
}

