% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BullsEyeR.R
\name{freqAnalysis}
\alias{freqAnalysis}
\title{Topic Modelling for content curation @COGNIZANT ANALYTICS}
\usage{
freqAnalysis(ds, spvar = 0.99, seedno = 12345, stemvar = 0)
}
\value{
See Description
}
\description{
1.freqAnalysis()- returns dataframe with words and their frequencies after initial preprocessing sparsity control and TFIDF analysis is performed.we can pick some words from the high frequency list as custom stop words.2.createDTM()- repeats first step, now including the custom stop words as well, removes empty documents if any and returns a Document term matrix. This DTM is used for finding optimal number of topics for LDA modelling  using FindTopicsNumber() from 'ldatuning' package3.BullsEye()- Performs preprocessing with custom stop words,topic number from ldatuning and builds unigram topic model with/without stemming. Returns    a)a data frame with keywords/dominant words of each document along with the topic number assigned by LDA model.b)a list of zero length documents after preprocessing c)a data frame with top 20 terms in all the topics discovered by LDA
}
