% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Sample text data}
\format{A csv file of text data}
\usage{
testdata
}
\description{
A collection of four articles with two columns- Article and Story namely.
}
\keyword{datasets}
