% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BullsEyeR.R
\name{freqAnalysis}
\alias{freqAnalysis}
\title{Functions
Frequency Analysis}
\usage{
freqAnalysis(ds, spvar = 0.99, stemvar = 0)
}
\arguments{
\item{ds}{a character vector of text documents}

\item{spvar}{a sparsity variable which defaults to 0.99}

\item{stemvar}{a variable indicating stemming to be performed or not which defaults to '0' meaning no stemming}
}
\value{
A dataframe with words and their frequencies after listed preprocessing.
}
\description{
The function freqAnalysis does a frequency analysis of retained words after initial preprocessing.
}
\examples{
\dontrun{
# Run it and see for yourself
}
data.tmp<-read.csv(system.file("ext", "testdata.csv", package="BullsEyeR"))
ds<-as.character(data.tmp$Story[1:2])
freqAnalysis(ds=ds,spvar=0.99,stemvar=0)
}
